chexit-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.24.0
tensorflow>=2.15.0
opencv-python-headless>=4.8.0
